+++ platform/core/src/services/DicomMetadataStore/DicomMetadataStore.ts
@@ -187,6 +187,10 @@ const BaseImplementation = {
     });
   },
   addSeriesMetadata(seriesSummaryMetadata, madeInClient = false) {
+    if (!seriesSummaryMetadata || !seriesSummaryMetadata.length || !seriesSummaryMetadata[0]) {
+      return;
+    }
+
     const { StudyInstanceUID } = seriesSummaryMetadata[0];
     let study = _getStudy(StudyInstanceUID);
     if (!study) {